# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https:\/\/app.solidtime.io\/api",
            "description": "Production"
        },
        {
            "url": "https:\/\/app.staging.solidtime.io\/api",
            "description": "Staging"
        },
        {
            "url": "https:\/\/soldtime.test\/api",
            "description": "Local"
        }
    ],
    "security": [
        {
            "oauth2": []
        }
    ],
    "paths": {
...
        "\/v1\/organization\/{organization}\/timesheet\/weeks": {
            "get": {
                "operationId": "getTimesheetWeeks",
                "summary": "Get timesheet weeks",
                "description": "Returns a list of weeks (most recent first) with total tracked seconds per week, used for the timesheet accordion layout.",
                "tags": [
                    "Timesheet"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of weeks to return (1-52)",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 52
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Number of weeks to skip",
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of week summaries",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#\/components\/schemas\/TimesheetWeekCollection"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#\/components\/responses\/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#\/components\/responses\/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#\/components\/responses\/ValidationException"
                    }
                }
            }
        },
        "\/v1\/organization\/{organization}\/timesheet": {
            "get": {
                "operationId": "getTimesheetGrid",
                "summary": "Get timesheet grid",
                "description": "Returns rows grouped by project+task with 7 day columns. Each cell contains total hours and related time entry IDs.",
                "tags": [
                    "Timesheet"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "week_start",
                        "in": "query",
                        "required": true,
                        "description": "Start date of the week (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "week_end",
                        "in": "query",
                        "required": true,
                        "description": "End date of the week (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Week grid data",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#\/components\/schemas\/TimesheetGridData"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#\/components\/responses\/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#\/components\/responses\/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#\/components\/responses\/ValidationException"
                    }
                }
            }
        },
        "\/v1\/organization\/{organization}\/timesheet\/cell": {
            "put": {
                "operationId": "updateTimesheetCell",
                "summary": "Update timesheet cell",
                "description": "Creates, updates, or deletes time entries for a given project+task+date cell. When hours > 0, creates or updates. When hours = 0, deletes all entries for that cell.",
                "tags": [
                    "Timesheet"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/TimesheetCellUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated cell data",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#\/components\/schemas\/TimesheetCellResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#\/components\/responses\/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#\/components\/responses\/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#\/components\/responses\/ValidationException"
                    }
                }
            }
        },
        "\/v1\/organization\/{organization}\/timesheet\/recent-tasks": {
            "get": {
                "operationId": "getTimesheetRecentTasks",
                "summary": "Get recent tasks",
                "description": "Returns recently used project+task combinations for the current user, ordered by most recently used.",
                "tags": [
                    "Timesheet"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of recent tasks to return (1-50)",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recent project+task combinations",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#\/components\/schemas\/TimesheetRecentTaskCollection"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#\/components\/responses\/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#\/components\/responses\/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#\/components\/responses\/ValidationException"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "https:\/\/solidtime.test\/oauth\/authorize"
                    }
                }
            }
        },
...
    }
}